combine task/template files
